Removed x86 ABI, added CLFS_CPU variable
Removed the x86 ABI page from being rendered in the book as it isn't
needed.  We weren't actually pick an ABI for x86, we were picking if
we wanted 32 or 64 bit versions of x86.  This would be better handled
if we were to pick which CPU architecture GCC should be compiled for
and then have a GCC that only outputs code for said 32 or 64 bit CPU.

Added the CLFS_CPU variable so that CPU architectures can be picked
from the list at:
http://gcc.gnu.org/onlinedocs/gcc/i386-and-x86_002d64-Options.html
We'll eventually use this CLFS_CPU choice with GCC's --with-arch=
configure line to generate a GCC just for our CPU.