Created MIPS and ARM GCC-Static files & cleanup
x86 no longer is common with MIPS and ARM for GCC-Static, so it has
its own GCC-Static file.  In the previous implementation, there
was a lot of duplication between the common and x86 GCC-Static files.
This removes the duplication by better referencing the common file
and also creates MIPS and ARM specific GCC-Static files.

In this way, everything that is common to all architectures goes into
the common file and each architecture's GCC-Static file is specific
to the architecture.

*** PLEASE NOTE ***: As of this commit, GCC-Final is BROKEN for x86!
It will be fixed shortly.
DO NOT MERGE THIS UNTIL GCC-FINAL IS FIXED FOR x86!