Moved chowning.xml to cleanup chapter, edited it
Moved bootable/common/chowning.xml to the cleanup chapter as that's a
more appropriate place for it.  Chowning everything to root should only
happen to the ${CLFS}-final directory (our copy) so as to not disrupt
the ability to go back and compile more packages as clfs user.

Also edited chowning.xml to only operate on the ${CLFS}-final directory
and to update chgrp command to use group "13" rather than "utmp" since
utmp group may not be group 13 on the build machine.