cross-tools/arm/abi: Provide recommendations on ABI
ABI selection for ARM is often confusing and the -support list often
has people reporting that they have difficulty selecting the best ABI
for their board.

Add some clarification and recommendations.  Move aapcs-linux and
apcs-gnu to the top of the table as those are really the only two
choices that should be used.

Signed-off-by: Andrew Bradford <andrew@bradfordembedded.com>