linux-headers: Install headers without a cp
The copy of headers from a dest/ dir to the final include directory is
due to 'make headers_install' deleting everything else in the
INSTALL_HDR_PATH directory before doing the install of the headers in
older kernels.  This seems to not be a problem any more, but regardless,
the kernel headers are the first thing installed into
${CLFS}/usr/include anyway so even if it did delete it's not a
problem.

This is 1 less step.